backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
openai==1.12.0
pydantic[email]==2.5.3
python-multipart==0.0.6
httpx==0.26.0
numpy==1.26.3
scikit-learn==1.4.0
sentence-transformers==2.3.1
Pillow==10.2.0
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
